Refactor decryption code into a package
