+++ go.mod
@@ -1,3 +1,3 @@
 module github.com/abustany/lcp-decrypt
 
-go 1.16
+go 1.22
